requirements.txt
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
nba_api>=1.4.1
supabase>=2.3.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
requests>=2.31.0
